% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fin_instr_autogen.R
\name{add.identifier.Data}
\alias{add.identifier.Data}
\title{add.identifier}
\usage{
\method{add.identifier}{Data}(this, ...)
}
\arguments{
\item{this}{Data}
}
\description{
add.identifier from FinancialInstrument library
}
