% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{addInstruments}
\alias{addInstruments}
\title{Add instruments to this by name, if it exists in Data.xlsx file}
\usage{
addInstruments(this, ...)
}
\arguments{
\item{this}{Data object}

\item{...}{character vector}
}
\value{
Data object
}
\description{
Add instruments to this by name, if it exists in Data.xlsx file
}
