% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fin_instr_autogen.R, R/generics.R
\name{buildHierarchy.Data}
\alias{buildHierarchy.Data}
\alias{buildHierarchy}
\title{buildHierarchy}
\usage{
\method{buildHierarchy}{Data}(this, ...)

buildHierarchy(this, ...)
}
\arguments{
\item{this}{Data object}

\item{...}{character, names of instruments and
character vector of fields from instruments}
}
\value{
data frame
}
\description{
buildHierarchy from FinancialInstrument library
}
