% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns_functions.R
\name{has.Sp}
\alias{has.Sp}
\alias{has.Di}
\alias{Sp}
\alias{Di}
\title{Check for actions in data}
\usage{
has.Sp(x, which = FALSE)

has.Di(x, which = FALSE)

Sp(x)

Di(x)
}
\arguments{
\item{x}{data object}

\item{which}{numeric, display position of match}
}
\description{
A set of functions to check for appropriate Dividends and Splits column names within a data object, as well as the availability and position of those columns.
}
