% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/modelData.R
\name{cloneData}
\alias{cloneData}
\alias{cloneData.Data}
\title{Make copy of object}
\usage{
cloneData(this)

\method{cloneData}{Data}(this)
}
\arguments{
\item{this}{Data object}
}
\value{
Data object
}
\description{
Make copy of object

Copy Data object
}
