% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fin_instr_autogen.R, R/generics.R
\name{currency.Data}
\alias{currency.Data}
\alias{currency}
\title{currency}
\usage{
\method{currency}{Data}(this, ...)

currency(this, ...)
}
\arguments{
\item{this}{Data}

\item{...}{character vector}
}
\value{
Data object
}
\description{
currency from FinancialInstrument library
}
