% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_from_list_xts}
\alias{data_from_list_xts}
\title{Create Data from list of xts}
\usage{
data_from_list_xts(
  l,
  names_from_list = TRUE,
  candles = FALSE,
  data = NULL,
  columns = c("Cl", "Ad", "Vo", "Di", "Sp")
)
}
\arguments{
\item{x}{xts}
}
\value{
Data
}
\description{
Create Data from list of xts
}
