% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{datastorage_options}
\alias{datastorage_options}
\title{Set or get option for datastorage package}
\usage{
datastorage_options(...)
}
\arguments{
\item{...}{option names to retrieve option values}
}
\value{

}
\description{
Set or get option for datastorage package
}
\section{Supported options}{

The following options are supported
\itemize{
  \item{\code{src}}
  \item{\code{currency}}
}
}

