% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fin_instr_fun.R, R/generics.R
\name{exchange_rate.Data}
\alias{exchange_rate.Data}
\alias{exchange_rate}
\title{Add exchange rate}
\usage{
\method{exchange_rate}{Data}(
  this,
  primary_id = NULL,
  currency = NULL,
  base_currency = NULL,
  tick_size = 0.01,
  identifiers = NULL,
  assign_i = TRUE,
  overwrite = TRUE,
  multiplier = 1,
  trade = TRUE,
  ...
)

exchange_rate(this, ...)
}
\arguments{
\item{this}{Data}

\item{primary_id}{character, name of instrument}

\item{currency}{character, counter currency for USDRUB it equals to RUB}

\item{base_currency}{character, base currency for USDRUB it equals to USD}

\item{tick_size}{numeric}

\item{identifiers}{list}

\item{assign_i}{logical}

\item{overwrite}{logical}

\item{multiplier}{numeric}

\item{...}{character vector}

\item{download}{character, name for downloading if it is not the same with primary_id}
}
\value{
Data object

Data object
}
\description{
For detailed description see FinancialInstrument::exchange_rate
}
