% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_to_counter.R
\name{exchange_to_counter}
\alias{exchange_to_counter}
\title{Exchange table in different currencies}
\usage{
exchange_to_counter(this, table, inds)
}
\arguments{
\item{this}{Data}

\item{table}{matrix with names from this object}

\item{inds}{numeric indexes}
}
\value{
matrix
}
\description{
Exchange table in different currencies
}
