% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fin_instr_fun.R, R/generics.R
\name{future.Data}
\alias{future.Data}
\alias{future}
\title{Add futures}
\usage{
\method{future}{Data}(this, ...)

future(this, ...)
}
\arguments{
\item{this}{Data object}

\item{...}{see in FinancialInstrument::future}
}
\value{
Data object

Data object
}
\description{
Add futures

Adds futures to this
}
