% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_subset}
\alias{get_subset}
\title{Get subset of Data object}
\usage{
get_subset(data, rows, columns, include_users_tables = TRUE)
}
\arguments{
\item{data}{Data}

\item{rows}{numeric vector}

\item{columns}{character vector}

\item{include_users_columns}{logical, whether get subset of user-defined tables or not}
}
\value{
Data
}
\description{
All tables will be subset
}
