% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelData.R
\name{getBuildPeriod.Data}
\alias{getBuildPeriod.Data}
\title{Returns period for building spread from storage}
\usage{
\method{getBuildPeriod}{Data}(this, recalc = FALSE)
}
\arguments{
\item{this}{Data}
}
\value{
character, period
}
\description{
Returns period for building spread from storage
}
