% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelData.R
\name{getDateByIndex.Data}
\alias{getDateByIndex.Data}
\title{Gets dates by indexes}
\usage{
\method{getDateByIndex}{Data}(this, indexes = NULL)
}
\arguments{
\item{this}{Data}

\item{indexes}{numeric vector}
}
\value{
vector of dates
}
\description{
Gets dates by indexes
}
