% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getDividends}
\alias{getDividends}
\title{Get dividends data from soures}
\usage{
getDividends(...)
}
\arguments{
\item{...}{params}

\item{Symbols}{character}

\item{split.adjust}{logical, default TRUE}
}
\value{
xts / data.frame
}
\description{
Get dividends data from soures
}
