% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fin_instr_autogen.R, R/generics.R
\name{getInstrument.Data}
\alias{getInstrument.Data}
\alias{getInstrument}
\title{getInstrument}
\usage{
\method{getInstrument}{Data}(this, ...)

getInstrument(this, ...)
}
\arguments{
\item{this}{Data}

\item{...}{character, instrument from this by id}
}
\value{
instrument
}
\description{
getInstrument from FinancialInstrument library
}
