% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getLast}
\alias{getLast}
\title{Get last point from soures}
\usage{
getLast(Symbols, ...)
}
\arguments{
\item{Symbols}{Data / character}

\item{...}{params}
}
\value{
xts / Data
}
\description{
Get last point from soures
}
