% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getLastCurrency}
\alias{getLastCurrency}
\title{Get last currency from finam and updates currency table}
\usage{
getLastCurrency(this)
}
\arguments{
\item{this}{Data}
}
\value{
Data
}
\description{
Get last currency from finam and updates currency table
}
