% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getSplits}
\alias{getSplits}
\title{Get splits data from soures}
\usage{
getSplits(...)
}
\arguments{
\item{...}{params}

\item{Symbols}{character}
}
\value{
xts / data.frame
}
\description{
Get splits data from soures
}
