% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSymbols_main.R
\name{getSymbols.Data}
\alias{getSymbols.Data}
\title{Get data for Data object}
\usage{
\method{getSymbols}{Data}(this, version = NULL, download_fun = download_instrument, ...)
}
\arguments{
\item{this}{Data}

\item{version}{character, heavy or light}

\item{...}{arguments for functions getSymbols_heavy and getSymbols_light}
}
\value{
Data
}
\description{
Get data for Data object
}
