% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Finam.R
\name{getSymbols.Finam_}
\alias{getSymbols.Finam_}
\title{Title}
\usage{
\method{getSymbols}{Finam_}(
  Symbols,
  env,
  return.class = "xts",
  index.class = "Date",
  from = "2007-01-01",
  to = Sys.Date(),
  adjust = FALSE,
  period = "day",
  col_funs = NULL,
  timecuts = list(),
  ...
)
}
\arguments{
\item{Symbols}{character vector}

\item{env}{environment, where assign value}

\item{return.class}{character, which class to return}

\item{index.class}{character, which index class to return}

\item{from}{character, start date}

\item{to}{character, end date}

\item{adjust}{bool, if TRUE then 00:00:00 - > 23:59:59}

\item{period}{character, one of 'day', '1min', '10min', 'hour'}

\item{...}{additional args}
}
\value{
nothing or table
}
\description{
Title
}
