% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getSymbols}
\alias{getSymbols}
\title{Get data from soures}
\usage{
getSymbols(Symbols, ...)
}
\arguments{
\item{Symbols}{Data / character}

\item{...}{params}
}
\value{
xts / Data
}
\description{
Get data from soures
}
