% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSymbols.R
\name{getSymbols.USDFIX}
\alias{getSymbols.USDFIX}
\title{Download RUBUSD from moex}
\usage{
\method{getSymbols}{USDFIX}(
  Symbols = "usdfix",
  from = Sys.Date() - 100,
  to = Sys.Date(),
  auto.assign = FALSE,
  ...
)
}
\arguments{
\item{Symbols}{here can be any character}

\item{from}{start date}

\item{to}{end date}

\item{auto.assign}{the same as in quantmod package}

\item{...}{additional params}
}
\value{
nothing or xts in according to auto.assign
}
\description{
Download RUBUSD from moex
}
