% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSymbols_new.R
\name{getSymbols_new}
\alias{getSymbols_new}
\title{Get data}
\usage{
getSymbols_new(this, prices = TRUE, tables = TRUE, ...)
}
\arguments{
\item{this}{Data}
}
\value{

}
\description{
This function returns lightweighted object.
}
