% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fin_instr_fun.R
\name{index.Data}
\alias{index.Data}
\title{Add stock}
\usage{
\method{index}{Data}(this, ...)
}
\arguments{
\item{this}{Data}

\item{...}{args to FinancialInstrument::stock}
}
\value{
Data object
}
\description{
Add stock
}
