% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fin_instr_autogen.R, R/generics.R
\name{instrument_attr.Data}
\alias{instrument_attr.Data}
\alias{instrument_attr}
\title{instrument_attr}
\usage{
\method{instrument_attr}{Data}(this, ...)

\method{instrument_attr}{Data}(this, ...)

instrument_attr(this, ...)
}
\arguments{
\item{this}{Data}

\item{...}{params}
}
\value{
Data object
}
\description{
instrument_attr from FinancialInstrument library

instrument_attr from FinancialInstrument library
}
