% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fin_instr_autogen.R, R/generics.R
\name{ls_currencies.Data}
\alias{ls_currencies.Data}
\alias{ls_currencies}
\title{ls_currencies}
\usage{
\method{ls_currencies}{Data}(this, ...)

ls_currencies(this)
}
\arguments{
\item{this}{Data}
}
\value{
Data object
}
\description{
ls_currencies from FinancialInstrument library
}
