% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fin_instr_autogen.R, R/generics.R
\name{ls_exchange_rates.Data}
\alias{ls_exchange_rates.Data}
\alias{ls_exchange_rates}
\title{ls_exchange_rates}
\usage{
\method{ls_exchange_rates}{Data}(this, ...)

ls_exchange_rates(this)
}
\arguments{
\item{this}{}
}
\value{
character vector, names of exchange rates in this
}
\description{
ls_exchange_rates from FinancialInstrument library
}
