% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fin_instr_autogen.R, R/generics.R
\name{ls_futures.Data}
\alias{ls_futures.Data}
\alias{ls_futures}
\title{ls_futures}
\usage{
\method{ls_futures}{Data}(this, ...)

ls_futures(this, ...)
}
\arguments{
\item{this}{Data}
}
\value{
character vector, futures' names in this
}
\description{
ls_futures from FinancialInstrument library
}
