% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/index.R
\name{ls_indexes}
\alias{ls_indexes}
\alias{ls_indexes.Data}
\title{Returns current currencies that this uses}
\usage{
ls_indexes(this, ...)

\method{ls_indexes}{Data}(this, ...)
}
\arguments{
\item{this}{Data}

\item{...}{params}
}
\value{
Data object
}
\description{
Returns current currencies that this uses

ls_indexes
}
