% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fin_instr_autogen.R, R/generics.R
\name{ls_instruments.Data}
\alias{ls_instruments.Data}
\alias{ls_instruments}
\title{ls_instruments}
\usage{
\method{ls_instruments}{Data}(this, ...)

ls_instruments(this)
}
\arguments{
\item{this}{Data}
}
\value{
character vector of instruments
}
\description{
ls_instruments from FinancialInstrument library
}
