% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fin_instr_autogen.R, R/generics.R
\name{ls_stocks.Data}
\alias{ls_stocks.Data}
\alias{ls_stocks}
\title{ls_stocks}
\usage{
\method{ls_stocks}{Data}(this, ...)

ls_stocks(this)
}
\arguments{
\item{this}{Data}
}
\value{
character vector of stocks
}
\description{
ls_stocks from FinancialInstrument library
}
