% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelData.R
\name{print.Data}
\alias{print.Data}
\title{print.Data}
\usage{
\method{print}{Data}(x, ...)
}
\arguments{
\item{x}{Data}

\item{...}{here can be additional arguments}
}
\value{
some text in console
}
\description{
print.Data
}
