% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{rm_currencies}
\alias{rm_currencies}
\title{Deletes specific currencies from this}
\usage{
rm_currencies(this, ...)
}
\arguments{
\item{this}{Data}

\item{...}{character vector of currencies}
}
\value{
nothing
}
\description{
Deletes specific currencies from this
}
