% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/modelData.R
\name{rm_data}
\alias{rm_data}
\alias{rm_data.Data}
\title{Delete all data from Data object}
\usage{
rm_data(this)

\method{rm_data}{Data}(this)
}
\arguments{
\item{this}{Data object}
}
\value{
nothing

nothing
}
\description{
Delete all data from Data object

Removes data from Data object
}
