% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fin_instr_autogen.R, R/generics.R
\name{rm_instruments.Data}
\alias{rm_instruments.Data}
\alias{rm_instruments}
\title{rm_instruments}
\usage{
\method{rm_instruments}{Data}(this, ...)

rm_instruments(this, ...)
}
\arguments{
\item{this}{Data object}

\item{...}{character vector, instruments to delete}
}
\value{
nothing
}
\description{
rm_instruments from FinancialInstrument library
}
