% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelData.R
\name{setBuildPeriod.Data}
\alias{setBuildPeriod.Data}
\title{Sets period for building spread from storage}
\usage{
\method{setBuildPeriod}{Data}(this, period)
}
\arguments{
\item{this}{Data}
}
\value{
Data
}
\description{
Sets period for building spread from storage
}
