% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSymbols_main.R
\name{setOrder.Data}
\alias{setOrder.Data}
\title{Set order of instruments}
\usage{
\method{setOrder}{Data}(this, x)
}
\arguments{
\item{this}{Data}

\item{x}{character, vector of primary ids}
}
\description{
Set order of instruments
}
