% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shrink_by_ind}
\alias{shrink_by_ind}
\title{Shrink Data}
\usage{
shrink_by_ind(data, start_ind, end_ind, copy = TRUE)
}
\arguments{
\item{data}{Data}

\item{start_ind}{numeric}

\item{end_ind}{numeric}

\item{copy}{logical}
}
\value{
Data
}
\description{
Shrink Data
}
