% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/modelData.R
\name{update_instruments}
\alias{update_instruments}
\alias{update_instruments.Data}
\title{Updates metadata of instruments from this}
\usage{
update_instruments(this, ...)

\method{update_instruments}{Data}(
  this,
  data,
  src = c("yahoo", "iShares", "masterDATA", "md", "morningstar", "SPDR", "TTR"),
  ...
)
}
\arguments{
\item{this}{Data object}

\item{...}{additional parameters, can be viewed in FinancialInstrument's methods}

\item{data}{character vector of instruments' names}

\item{src}{character one of c("yahoo","iShares","masterDATA","md","morningstar","SPDR","TTR")}
}
\value{
Data
}
\description{
Updates metadata of instruments from this

Updates instruments metadata
}
