% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/updateSymbols.R
\name{updateLast}
\alias{updateLast}
\alias{updateLast.Data}
\title{Update last point in Data object}
\usage{
updateLast(this)

\method{updateLast}{Data}(this)
}
\arguments{
\item{this}{Data}
}
\value{
Data object


}
\description{
Update last point in Data object

Update the last point
}
