% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/updateSymbols.R
\name{updateSymbols}
\alias{updateSymbols}
\alias{updateSymbols.Data}
\title{Update Data object}
\usage{
updateSymbols(this)

\method{updateSymbols}{Data}(this)
}
\arguments{
\item{this}{Data}
}
\value{
Data object


}
\description{
Update Data object

Update all tables to Sys.time()
}
