% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fin_instr_autogen.R, R/generics.R
\name{add.identifier.Data}
\alias{add.identifier.Data}
\alias{add.identifier}
\title{add.identifier}
\usage{
\method{add.identifier}{Data}(this, ...)

add.identifier(this, ...)
}
\arguments{
\item{this}{Data}

\item{...}{params}
}
\value{
Data object
}
\description{
add.identifier from FinancialInstrument library
}
