% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fin_instr_autogen.R
\name{instrument_attr.Data}
\alias{instrument_attr.Data}
\title{instrument_attr}
\usage{
\method{instrument_attr}{Data}(this, ...)
}
\arguments{
\item{this}{Data}
}
\description{
instrument_attr from FinancialInstrument library
}
