% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/modelData.R
\name{modify}
\alias{modify}
\alias{modify.Data}
\title{Lets change arguments of this}
\usage{
modify(this, ...)

\method{modify}{Data}(this, ...)
}
\arguments{
\item{this}{Data}

\item{...}{one of c("period", "currency", "freq", "from", "to", 'period_build', 'columns', 'candles')}
}
\value{
Data object

Data object
}
\description{
Lets change arguments of this

Change settings in Data
}
