% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fin_instr_fun.R, R/generics.R
\name{stock.Data}
\alias{stock.Data}
\alias{stock}
\title{Add stock}
\usage{
\method{stock}{Data}(this, ...)

stock(this, ...)
}
\arguments{
\item{this}{Data}

\item{...}{character vector of stocks}
}
\value{
Data object

Data object
}
\description{
Add stock

Adds stock to this
}
